+++ app/src/main/java/com/maalelan/postcardstorehouse/fragments/AddPostcardFragment.java
@@ -1,10 +1,24 @@
 package com.maalelan.postcardstorehouse.fragments;
 
 import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.maalelan.postcardstorehouse.R;
 
 public class AddPostcardFragment extends Fragment {
 
-
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        // Inflate the layout for this fragment and return the root view
+        return inflater.inflate(R.layout.add_postcard_fragment, container, false);
+    }
 }