+++ documentation/changelog.md
@@ -18,17 +18,17 @@ Types of changes for quick reminder.
 ### Added
 - Initial project structure defined
 - `README.md` created with project description and folder overview
+- `troubleshooting.md` created with first issues
 - Postcard data model (`Postcard.java`) created
 - PostcardImage data model (`PostcardImage.java`) created
 - Room database setup: `PostcardEntity`, `PostcardDao`, and `PostcardDatabase`
 - Room database .fallbackToDestructiveMigration() is deprecated 
   - -> migrations dir: `migration1to2`, for later use
 - Bottom navigation + UI fragments
-  - README.md updated
 - Fragments setup: `AddPostcardFragment`,`GalleryFragment`, `HomeFragment`,`PostcardDetailsFragment`
 - Res/navigation nav_graph.xml
 - Res/menu bottom_nav_menu.xml
-- Res/layout: home_fragment.xml
+- Res/layout: home_fragment.xml, add_postcard_fragment.xml
 - Drawable/ some icons, background image
 - Dependencies
   - Navigation